Rewrite AI from scratch
* Now it tries to define what the axis of the attacked ship is
(vertical or horizonal), and attacks in this axis.
Instead of attacking a random cell relative to the first hit at ship,
hits, it now guesses the most probable options where a ship may be,
compares it with valid attacks (the cells that are not yet attacked),
and attacks. It is capable of determining the direction of the attack,
so that it won't draw a solid line of attacks with length 10 when all
it needs is to attack a ship with length 1 to 4.
* Properties that define attack are now private.
* Apart from that, I added getters and setters for some of these
properties so that tests can make sure pure functions are working
correctly